# Conflicts:
#   BE/gumid101/src/main/java/com/ssafy/gumid101/crew/CrewManagerRepository.java
#   BE/gumid101/src/main/java/com/ssafy/gumid101/crew/CrewManagerRestController.java
#   BE/gumid101/src/main/java/com/ssafy/gumid101/crew/CrewManagerService.java
#   BE/gumid101/src/main/java/com/ssafy/gumid101/crew/CrewManagerServiceImpl.java
#   BE/gumid101/src/main/java/com/ssafy/gumid101/crew/UserCrewJoinRepository.java

같은 구역을 기능만 따져서 업무를 나누다 보니, 충돌이 많이 생겼다.

+++ BE/gumid101/src/main/java/com/ssafy/gumid101/dto/RecruitmentParamsDto.java
@@ -17,18 +17,18 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class RecruitmentParamsDto implements Serializable {
-	private int size;
-	private int offset;
+	private Integer size;
+	private Integer offset;
 	private String title;
 	private String startDay;
 	private String endDay;
 	private String startTime;
 	private String endTime;
-	private int pointMin;
-	private int pointMax;
+	private Integer pointMin;
+	private Integer pointMax;
 	private String purposeType;
-	private double purposeMinValue;
-	private double purposeMaxValue;
-	private int goalMinDays;
-	private int goalMaxDay;
+	private Integer purposeMinValue;
+	private Integer purposeMaxValue;
+	private Integer goalMinDays;
+	private Integer goalMaxDay;
 }
